Add from_pandas function to top level module
Per discussion in https://github.com/great-expectations/great_expectations/issues/225
this adds a top level method from_pandas to allow for the direct
construction of a ge dataframe from a pandas dataframe. This
is effectively equivalent to ge.dataset.PandasDataSet(df), but is more
discoverable and has an explicit expectations_config and goes
through the same codepath as the other top-level construction functions
read_csv and read_json.