Add details to CONTRIBUTING
Some fairly big changes to the contribution instructions:
- restructured the sections
- added instructions for docs-only contribution
- added step-by-step instructions for code contributions including forks etc

I tested these with a new github account and I'm pretty confident they make sense, but there's definitely a chance I missed a step. I anticipate iterations on this for both clarity and correctness, but this should be a good starting point for more detailed instructions.