Refactor branch name master->main.
This refactor to eliminate the branch name 'master' is a way for us recognize that discrimination is a subtle but pervasive part of our cultural heritage. It is a small but concrete step that symbolizes our resolution to act in ways that empower all of our friends, neighbors, and fellow citizens to fully participate in and enjoy the benefits of our community.

Signed-off-by: James Campbell <james.p.campbell@gmail.com>