Update the SuiteEditNotebookRenderer.add_header (#1892)
Pass environment variables to render_with_overwrite method to allow using environment variables in the jinja templates. This allows use cases where depending on environment variables the notebook code generation can change. 

We have patched this on our side so we can easily run a profiler on a local/in-memory pandas dataset and then automatically call the `great_expectations suite edit` command and already have the dataset loaded in pandas (which is otherwise not supported since GE does know where the data lives in this case).

Co-authored-by: Eugene Mandel <eugene.mandel@gmail.com>