test(prof): Add taxi data, script, & analysis
This commit contains the following:
- taxi_yellow_trip_data_samples: Uniform samples from the original datasets
- taxi_sampling_script: What was used to generate the above samples
- taxi_sampling_stats: Analyzes basic metrics found in the data

The data should include everything from 2018 - 2020 (with the first 3
months of 2018 coming from prior commits). Each CSV contains 10,000 rows
and equivalent column titles.

Note that there are some rows that contain negative or NaN values; these
are noted in the Jupyter Notebook but may play a role when using this
data for unit/integration tests.