Update snippet line range in tutorial
In the section of the tutorial "Validate data" (https://docs.greatexpectations.io/docs/tutorials/getting_started/tutorial_validate_data), the snippet displayed for configuring a checkpoint has extra lines on the top and missing lines in the bottom.
When used, it makes the execution fail. Some lines were probably added above in the file, hence the shift in what is actually displayed.