[BUGFIX] Change GXSqlDialect.AWSATHENA to awsathena
The SQL dialect in SQLalchemy for Athena is awsathena.
This line was changed on commit b2ec0c8c6bdb99475f80e6987d40de8cd230bf6f

But with the current configuration profiling does not work with athena,
it fails when trying to execute percentiles.
By changing the AWSATHENA constant to "awsathena" we can fix that issue
with profiling