Add SQL date time format expectation
expect_column_values_to_match_sql_date_time_format is similar to
expect_column_values_to_match_strftime_format but uses SQL date
time patterns, e.g.
https://spark.apache.org/docs/latest/sql-ref-datetime-pattern.html
.

The motivation for this comes from
https://github.com/great-expectations/great_expectations/issues/8929
. It is hard to natively implement the strftime check in Spark and
SQL, whereas it is much more natural to use the SQL date time
patterns. Using the SQL date time check would avoid this issue
with the query referencing an internal function.

It is currently only implemented for Spark. I believe it should be
relatively straightforward to implement for SQL but, for the same
reasons it's difficult to implement the strftime approach for SQL,
it will not be straightforward to implement this for pandas.

NOTE: in contrast to the strftime approach, the SQL date time
format check succeeds when the column values are already in a
date time format.